plugin bundle content payloads should now be arefs
Bundle config looks like [ name, pkg, payload ].  Previously, payload was a
hashref, and arrayref values became multivalue.  This was a bizarre and
unfortunate decision.  It should have mirrored the conceptual model, which is a
sequence of key/value pairs.  Arrayref in this form are now accepted, and
hashref payloads will be removed at some unspecified future date.